requirements.txt
langchain
langchain-community
faiss-cpu
streamlit
python-dotenv
pypdf
sentence-transformers
groq
